Allow reporting from haproxy socket.
Adding a new configuration option - haproxy_socket - which
overrides the haproxy_url setting.
When set to the path of haproxy stat socket, will use this
socket as a source of stats.
In addition to statistics also available on the CSV interface,
it reports general process informations (under the metric
path : .general_process_info)